sign check

+++ src/main/java/com/tapp/api/v1/controllers/StickerController.java
@@ -1,11 +1,17 @@
 package com.tapp.api.v1.controllers;
 
+import com.tapp.api.v1.exceptions.SignCheckException;
 import com.tapp.api.v1.models.Sticker;
+import com.tapp.api.v1.models.User;
 import com.tapp.api.v1.services.StickerService;
+import com.tapp.api.v1.services.UserService;
+import com.tapp.api.v1.utils.ParamsUtil;
+import com.tapp.api.v1.utils.UserRoles;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.MalformedURLException;
 import java.util.List;
-
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
 @RestController
@@ -13,6 +19,7 @@ import java.util.concurrent.ExecutionException;
 @RequestMapping("v1/stickers")
 public class StickerController {
     private StickerService stickerService = new StickerService();
+    private UserService userService = new UserService();
 
     @GetMapping
     List<Sticker> getSticker() {
@@ -25,18 +32,63 @@ public class StickerController {
     }
 
     @PostMapping
-    String addSticker(@RequestBody Sticker sticker) throws ExecutionException, InterruptedException {
-        long id = stickerService.addSticker(sticker).get();
-        return "{'id': "+ id +" }";
+    CompletableFuture<Sticker> addSticker(@RequestHeader("params") String params,
+                                          @RequestBody Sticker sticker) {
+        try {
+            if (ParamsUtil.isAuthentic(params)) {
+                User user = userService.getUser(ParamsUtil.getUserId(params)).get();
+                if (user.getRole().equals(UserRoles.admin.toString())) {
+                    return stickerService.addSticker(sticker);
+                }
+            }
+            throw new UnsupportedOperationException();
+
+        } catch (SignCheckException e) {
+            throw new UnsupportedOperationException();
+        } catch (MalformedURLException | InterruptedException | ExecutionException e) {
+            e.printStackTrace();
+            throw new UnsupportedOperationException();
+        }
+
     }
 
     @PatchMapping
-    void updateSticker(@RequestBody Sticker sticker) {
-        stickerService.updateSticker(sticker);
+    void updateSticker(@RequestHeader("params") String params,
+                       @RequestBody Sticker sticker) {
+        try {
+            if (ParamsUtil.isAuthentic(params)) {
+                User user = userService.getUser(ParamsUtil.getUserId(params)).get();
+                if (user.getRole().equals(UserRoles.admin.toString())) {
+                    stickerService.updateSticker(sticker);
+                }
+            }
+            throw new UnsupportedOperationException();
+
+        } catch (SignCheckException e) {
+            throw new UnsupportedOperationException();
+        } catch (MalformedURLException | InterruptedException | ExecutionException e) {
+            e.printStackTrace();
+            throw new UnsupportedOperationException();
+        }
     }
 
     @DeleteMapping("{id}")
-    void deleteSticker(@PathVariable long id) {
-        stickerService.deleteSticker(id);
+    void deleteSticker(@RequestHeader("params") String params,
+                       @PathVariable long id) {
+        try {
+            if (ParamsUtil.isAuthentic(params)) {
+                User user = userService.getUser(ParamsUtil.getUserId(params)).get();
+                if (user.getRole().equals(UserRoles.admin.toString())) {
+                    stickerService.deleteSticker(id);
+                }
+            }
+            throw new UnsupportedOperationException();
+
+        } catch (SignCheckException e) {
+            throw new UnsupportedOperationException();
+        } catch (MalformedURLException | InterruptedException | ExecutionException e) {
+            e.printStackTrace();
+            throw new UnsupportedOperationException();
+        }
     }
 }
